Refactors the callbacks and random selection
